backend/requirements.txt
asgiref==3.11.1
cloudinary==1.44.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
dj-database-url==3.0.1
gunicorn==23.0.0
pillow==12.1.1
psycopg[binary]==3.2.12
python-dotenv==1.0.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
